eliminating top-level build.gradle file, simplifying project configuration and adding missing license information to files
