+++ gradle.properties
@@ -2,6 +2,10 @@ VERSION_NAME=1.4.0
 VERSION_CODE=7
 GROUP=com.github.pwittchen.kirai
 
+POM_NAME=kirai
+POM_ARTIFACT_ID=library
+POM_PACKAGING=jar
+
 POM_DESCRIPTION=Kirai is flavored string formatting library
 POM_URL=https://github.com/pwittchen/kirai
 POM_SCM_URL=https://github.com/pwittchen/kirai
