--- library/build.gradle
@@ -1,45 +1,0 @@
-apply plugin: 'java'
-apply plugin: 'idea'
-apply plugin: 'jacoco'
-
-sourceCompatibility = 1.7
-targetCompatibility = 1.7
-
-repositories {
-    mavenCentral()
-    jcenter()
-}
-
-dependencies {
-    testCompile 'junit:junit:4.12'
-    testCompile 'com.google.truth:truth:0.25'
-}
-
-sourceSets {
-    main {
-        java {
-            srcDirs 'src/main/java'
-        }
-    }
-
-    test {
-        java {
-            srcDir 'src/test/java'
-        }
-    }
-}
-
-jacocoTestReport {
-    reports {
-        xml.enabled = true
-        html.enabled = true
-    }
-}
-
-check.dependsOn jacocoTestReport
-
-task wrapper(type: Wrapper) {
-    gradleVersion = '2.2.1'
-}
-
-apply from: '../maven_push.gradle'
