--- library/gradle.properties
@@ -1,3 +1,0 @@
-POM_NAME=kirai
-POM_ARTIFACT_ID=library
-POM_PACKAGING=jar
