+++ src/main/java/com/github/pwittchen/kirai/library/Piece.java
