+++ src/main/java/com/github/pwittchen/kirai/library/Preconditions.java
