+++ src/main/java/com/github/pwittchen/kirai/library/Syntax.java
