+++ src/main/java/com/github/pwittchen/kirai/library/html/HtmlPiece.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright (C) 2015 Piotr Wittchen
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.pwittchen.kirai.library.html;
+
+import com.github.pwittchen.kirai.library.Piece;
+
+public final class HtmlPiece extends Piece {
+  private HtmlPiece(String key, Object value) {
+    super(key, value, new HtmlSyntax());
+  }
+
+  public static Piece put(String key, Object value) {
+    return new HtmlPiece(key, value);
+  }
+}
