+++ src/main/java/com/github/pwittchen/kirai/library/terminal/TerminalColor.java
