+++ src/test/java/com/github/pwittchen/kirai/library/TerminalPieceTest.java
