+++ src/test/java/com/github/pwittchen/kirai/library/TerminalSyntaxTest.java
